pyproject.toml
[project]
name = "tech-news-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "feedparser>=6.0.12",
    "html5lib>=1.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "lxml>=6.0.2",
    "nest-asyncio>=1.6.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
]
